+++ sharding-orchestration/sharding-orchestration-jdbc/src/test/java/io/shardingsphere/jdbc/orchestration/api/yaml/fixture/DecrementKeyGenerator.java
@@ -15,7 +15,7 @@
  * </p>
  */
 
-package io.shardingsphere.jdbc.orchestration.yaml.fixture;
+package io.shardingsphere.jdbc.orchestration.api.yaml.fixture;
 
 import io.shardingsphere.core.keygen.KeyGenerator;
