+++ sharding-orchestration/sharding-orchestration-jdbc/src/test/java/io/shardingsphere/jdbc/orchestration/api/yaml/fixture/MultiAlgorithm.java
@@ -15,18 +15,17 @@
  * </p>
  */
 
-package io.shardingsphere.jdbc.orchestration.yaml.fixture;
+package io.shardingsphere.jdbc.orchestration.api.yaml.fixture;
 
-import io.shardingsphere.core.keygen.KeyGenerator;
+import io.shardingsphere.core.api.algorithm.sharding.ShardingValue;
+import io.shardingsphere.core.api.algorithm.sharding.complex.ComplexKeysShardingAlgorithm;
 
-import java.util.concurrent.atomic.AtomicInteger;
+import java.util.Collection;
 
-public final class IncrementKeyGenerator implements KeyGenerator {
-    
-    private static final AtomicInteger SEQUENCE = new AtomicInteger(100);
+public final class MultiAlgorithm implements ComplexKeysShardingAlgorithm {
     
     @Override
-    public Number generateKey() {
-        return SEQUENCE.incrementAndGet();
+    public Collection<String> doSharding(final Collection<String> availableTargetNames, final Collection<ShardingValue> shardingValues) {
+        return null;
     }
 }
