+++ sharding-orchestration/sharding-orchestration-jdbc/src/test/resources/logback-test.xml
@@ -1,0 +1,18 @@
+<?xml version="1.0"?>
+<configuration>
+    <appender name="console" class="ch.qos.logback.core.ConsoleAppender">
+        <encoder>
+            <pattern>[%-5level] %d{HH:mm:ss.SSS} [%thread] %logger{36} - %msg%n</pattern>
+        </encoder>
+    </appender>
+    <logger name="io.shardingsphere" level="warn" additivity="false">
+        <appender-ref ref="console"/>
+    </logger>
+    <logger name="io.shardingsphere.jdbc.orchestration.reg.newzk.client.retry.RetryCallable" level="error" additivity="false">
+        <appender-ref ref="console"/>
+    </logger>
+    <root>
+        <level value="error" />
+        <appender-ref ref="console" />
+    </root>
+</configuration>
