+++ sharding-orchestration/sharding-orchestration-reg/src/test/java/io/shardingsphere/orchestration/reg/newzk/client/cache/PathResolveTest.java
@@ -15,7 +15,7 @@
  * </p>
  */
 
-package io.shardingsphere.jdbc.orchestration.reg.newzk.client.cache;
+package io.shardingsphere.orchestration.reg.newzk.client.cache;
 
 import org.junit.Test;
